drivers: phy: dp83867: add handling for RX strap quirk during PHY init
Add the handling of the optional boolean device tree flag
"ti,dp83867-rxctrl-strap-quirk" to the driver.

Whenever this flag is set, the INT_TST_MODE_1 bit [7] in the CFG4
register within the PHY's external register set is cleared explicitly
during the initalization of the PHY, preventing the PHY from getting
stuck in self-test mode after power up with a certain bootstrap pin
setup.

Signed-off-by: Immo Birnbaum <mail@birnbaum.immo>